Developper
we are looking for a developer for our server! we are also looking for Staff police preparedness and much more you are one of those who can develop of their own free will so now you have to strike and contact us on the discord! we are waiting for you